It's still broken

+++ pollstar/src/main/java/com/Toine/pollstar/Repository/JPARepository/JPA/IChoiceRepositoryJPA.java
@@ -4,9 +4,10 @@ import com.Toine.pollstar.Core.Model.Choice;
 import com.Toine.pollstar.Core.Model.Poll;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.ArrayList;
 import java.util.List;
 
-public interface IChoiceRepositoryJPA extends JpaRepository<Choice, Long>
+public interface IChoiceRepositoryJPA extends JpaRepository<Choice, Integer>
 {
     //List<Choice> getAllByPollPollID(Poll poll);
     List<Choice> getChoiceByPoll(Poll poll);
